move minisocket.js
# Conflicts:
#	cli/common/bsb.js
