+++ cli/common/minisocket.js
